\name{bandit-package}
\alias{bandit-package}
\alias{bandit}
\docType{package}
\title{Functions for simple A/B split test and multi-armed bandit analysis}
\description{A set of functions for doing analysis of A/B split test data and web metrics in general.}
\details{
  \tabular{ll}{
    Package: \tab bandit\cr
    Type: \tab Package\cr
    Title: \tab Functions for simple A/B split test and multi-armed bandit analysis\cr
    Version: \tab 0.5.0\cr
    Date: \tab 2014-05-03\cr
    Imports: \tab boot, gam (>= 1.09)\cr
    Author: \tab Thomas Lotze and Markus Loecher\cr
    Maintainer: \tab Thomas Lotze <thomaslotze@thomaslotze.com>\cr
    License: \tab GPL-3\cr
  }
}
\author{Thomas Lotze and Markus Loecher}

\keyword{ package }
