\name{bandit-package}
\alias{bandit-package}
\alias{bandit}
\docType{package}
\title{
Functions for simple A/B split test and multi-armed bandit analysis
}
\description{
A set of functions (intended to be used by the ruby gem metricizer) for doing basic analysis of A/B split test data and recommending distribution of next probabilities, with some additional convenience functions for web metrics analysis.
}
\details{
\tabular{ll}{
Package: \tab bandit\cr
Type: \tab Package\cr
Version: \tab 0.4.1\cr
Date: \tab 2012-07-07\cr
Depends: \tab boot\cr
License: \tab GPL-3\cr
}
}
\author{
Thomas Lotze <thomaslotze@thomaslotze.com>
}
\keyword{ package }
